[GH workflows] Update clean_gh_releases.py to purge after 30 days (#186)
Keeping old nightlies around for a bit longer is friendlier for other projects, e.g. in terms of PRs (e.g. on https://github.com/llvm/lighthouse) that are open for a longer period and still have the old nightly in their environment config files.

Torch-mlir is looking like it will switch to keeping 60 days of old nightlies: https://github.com/llvm/torch-mlir-release/pull/21 . For eudsl, I come to a limit of roughly 30 days due to every nightly consisting of 30 assets or less (the mlir-python-bindings being at 30 a night). With the purge after 30 days, it should be we don't run afoul of GitHub's 1000 assets-per-release limit.