[eudsl-python-extras] drop support for generics for py < 3.12 (#236)
After much consternation and flaggelation, I've decided we
can't/shouldn't support generics for python < 3.12. The context/reason
is that I thought I had robust support for just plain `TypeVar`s as
globals (and in the closure) but recently I've realized that that
support is not robust.

In the weeds: the "generic" kernel instantiation effectively mutates the
`TypeVar` and thus when the `TypeVar` is a global, multiple
instantiations of the same kernel (with different concrete params) will
race (on that global `TypeVar`). `TypeVar`s in the closure don't suffer
from this because we copy the `body_builder` and its closure when
instantiating the kernel. In 3.12 (when using the brackets `def fun[M,
N, ...]` syntax) the `TypeVar`s are always in the closure[^1]. I tried
to figure out somehow to move globals into closures but failed and
decided it's not worth the effort anyway.

Note, the added test (`test_generics`) is gated behind 3.13 because
generics default value assignment is a 3.13 feature and the 3.12 parser
will choke on it, but everything else works in 3.12.

[^1]: That's actually a CPython implementation detail - there's a
compiler transform which rewrites the function into a closed over
function with those TypeVars in its closure.