[mlir-python-bindings-wasm] add `WasmExecutionEngine` (#248)
This PR adds runtime functionality for the mlir-python-bindings-wasm
wheels. The way it works is roughly the same as
[clang-interpreter](http://github.com/llvm/llvm-project/blob/a941e150749650e6a75e948f10d46b0bedcc128b/clang/lib/Interpreter/Wasm.cpp#L120-L126):
the MLIR is compiled to wasm, fully linked into a shared object, and
then `dlopen`ed (function symbols are accessed using `dlsym`).

A couple of things to call out:

1. We're using the `wasm32` target so lowering to LLVM requires using
`index-bitwidth=32` in various passes;
2. The various `numpy` -> `memref_descriptor` helpers need to be
adjusted to work with 32b instead of 64b;
3. For some reason (🤷) `@main` is not a valid function name (collides
with something somewhere in the emscripten runtime).

Follow-up PRs will extend `LLVMBackend` in python-extras to work with
this.