Add eudsl-python-extras packages to Wasm playgrounds (#266)
This is a follow-up PR of #264.

We try to add eudsl-python-extras packages to these two Wasm
playgrounds. Currently the released packages of eudsl-python-extras are
sdist which is not quite compatible to `pyodide.loadPackage` and
`piplite`, so we first pack it as a wheel package via `pip wheel`, and
then pack these wheel packages into the playgrounds.

In JupyterLite, we can install the package via:
```python
import piplite
await piplite.install("eudsl-python-extras")
```

Preview changes at:
- https://llvm.github.io/eudsl/jupyter/ 
- https://llvm.github.io/eudsl/console/